% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Genetic Operator: Chromosome Crossover and Mutation}
\usage{
crossover(chr1, chr2)
}
\arguments{
\item{chr1, chr2}{a numeric vectors represents individual parent chromosome.}
}
\value{
A C by 2 matrix with each column representing the offspring from a genetic
operator of crossover and mutation, C is the chromosome length.
}
\description{
Make 2 individual parent chromosomes crossover and
mutate when breeding offsprings.
}
\details{
This function makes two individual parent chromosomes perfrom
crossover and mutation when breeding offspring.
Note that the crossover is simply one-point crossover and the mutation
is based on \code{GA::mutation()}.
}
\examples{
ind1 <- initialization(10, 15)[, 1]
ind2 <- initialization(10, 15)[, 2]
crossover(ind1, ind2)
}
