% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialization}
\alias{initialization}
\title{Generate the Initial Generation Randomly}
\usage{
initialization(C, P)
}
\arguments{
\item{C}{chromosome length (number of predictor variables)}
}
\value{
A Bolean Matrix with dimension C by P where each column
representing a chromosome, in which T marks that the gene(variable)
as active and F as inactive.
}
\description{
Bernoulli sampling to form the initial generation of the
Genetic Algorithm.
}
\details{
This function produces initial generation given the chromosome
length for Genetic Algorithm and the population size. The row of the
generated boolean matrix represents the locus of a gene(variable), and the
column of the matrix represents different members of the first generation.
}
