% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutation}
\alias{mutation}
\title{Chromosome Mutation}
\usage{
mutation(chr)
}
\arguments{
\item{chr}{a boolean vector representing an individual chromosome.}
}
\value{
Return a mutated chromosome vector with the same length as input.
}
\description{
Make a chromosome with fixed probability 0.01 to mutate.
}
\details{
This function makes a chromosome have a 1% fixed chance to mutate
in each locale. If mutation happens at one locale, it will make the value
in that locale from T to F or from F to T.
}
