% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextGeneration.R
\name{nextGeneration}
\alias{nextGeneration}
\title{Breed the Selected Parents Generated From \code{GA::selection()}}
\usage{
nextGeneration(pop, selResult, G)
}
\arguments{
\item{pop}{boleans matrix determined by \code{GA::initialization()}}

\item{selResult}{list returnd by \code{GA::selection()}
At the minimum, it needs to contain indices for the selected parents 1 and 2.}

\item{offspringNum}{number of offspring generated to update the generation.}
}
\value{
Returns a C by P maatrix containing the population for the next
generation.
}
\description{
Breed the Selected Parents Generated From \code{GA::selection()}
}
\details{
Breeding uses \code{GA::crossover()} for each pair of parents. The
Generation Gap G is the proportion of the generation to be replaced by
generated offspring. If the number of the selected population produced by
generation gap is not an even number, choose the nearest bigger even number.
Uses the offspringNum number of offspring to replace the least fit
individuals in parent generation
}
\examples{
dat <- mtcars
C <- ncol(dat) - 1
P <- as.integer(1.5 * C)
pop <- initialization(C, P)
fitnessFunction <- AIC
model <- lm
y <- as.matrix(dat[1])
X <- as.matrix(dat[-1])
fitScores <- fitness(pop, y, X, fitnessFunction, model, dat)
offspringNum <- 4
selResult <- selection(pop, fitnessFunction, model, fitScores, offspringNum, method = 1, dat, K = 2)
nextGeneration(pop, selResult, offspringNum)
}
