% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGA.R
\name{plotGA}
\alias{plotGA}
\title{Plotting the Progress of the Genetic Algorithm}
\usage{
plotGA(results)
}
\description{
Make a chromosome with fixed probability 0.01 to mutate.
}
\details{
This function plots the output list from \code{GA::select()}. The plot shows fitness
scores of each model in each generation as well as the best fitting model (in green).
}
\examples{
results <- select(mtcars)
plotGA(results)
}
